requirements.txt
django >= 1.5
django-debug-toolbar
south
django-mptt
python-dateutil
sorl-thumbnail